requirements.txt
python-dotenv==0.19.2
openai==0.27.0
